[MINOR][UI] Escape HTML when rendering error messages in the Web UI tables
Task error messages and stage failure reasons are rendered by the Web UI
JavaScript (the stage-page task table, and app names on the history page)
by concatenating the text directly into HTML. Error text that contains
HTML metacharacters (for example generic type signatures such as Seq<Int>,
or other angle brackets) is therefore rendered incorrectly and can break
the surrounding table markup.

Escape such text consistently with the existing server-side scala.xml
rendering, and add a regression test.
